+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added validation for repos hosting the organization site, i.e. <org-id>.github.io. ([#83](https://gitlab.eclipse.org/eclipsefdn/security/otterdog/-/issues/83))
 - Added validation for secrets and webhooks to issue a warning if a value is provided that does not use a credential provider.
 - Added operation `delete-file` to delete files in a repo of an organization.
 - Added support for workflow settings for an organization. ([#62](https://gitlab.eclipse.org/eclipsefdn/security/otterdog/-/issues/62))